Various updates
Adds documentation for compression and credits to documentation intro. Adds bases 11-15 to the compressor for compressing integer arrays. Replaces the documentation menus with a single menu. Adds the ability to use customised markdown in HTML docs. Also adds a (not yet functional) settings section to dcoumentation which will allow for the reordering of documentation sections. Fixes various bugs.